migrated cac tests to convert command
