+++ pkg/cmd/project/create/create_test.go
@@ -1,10 +1,7 @@
 package create_test
 
 import (
-	"net/url"
 	"testing"
-
-	"github.com/OctopusDeploy/go-octopusdeploy/v2/pkg/credentials"
 
 	"github.com/OctopusDeploy/cli/pkg/cmd"
 	projectCreate "github.com/OctopusDeploy/cli/pkg/cmd/project/create"
@@ -15,10 +12,6 @@ import (
 	"github.com/OctopusDeploy/go-octopusdeploy/v2/pkg/projectgroups"
 	"github.com/stretchr/testify/assert"
 )
-
-var serverUrl, _ = url.Parse("https://serverurl")
-var spinner = &testutil.FakeSpinner{}
-var rootResource = testutil.NewRootResource()
 
 func TestAskProjectGroup_WithProvidedName(t *testing.T) {
 	value, _, err := projectCreate.AskProjectGroups(nil, "FooBar", nil, nil)
@@ -96,146 +89,17 @@ func TestPromptForConfigAsCode_NotUsingCac(t *testing.T) {
 	asker, checkRemainingPrompts := testutil.NewMockAsker(t, pa)
 	flags := projectCreate.NewCreateFlags()
 	flags.ConfigAsCode.Value = false
+
+	var convertCallbackCalled bool
 	opts := projectCreate.NewCreateOptions(flags, &cmd.Dependencies{Ask: asker})
-	err := projectCreate.PromptForConfigAsCode(opts, nil)
+	opts.ConvertProjectCallback = func() (cmd.Dependable, error) {
+		convertCallbackCalled = true
+		return nil, nil
+	}
+
+	_, err := projectCreate.PromptForConfigAsCode(opts)
 	checkRemainingPrompts()
 	assert.NoError(t, err)
+	assert.False(t, convertCallbackCalled)
 	assert.False(t, opts.ConfigAsCode.Value)
-}
-
-func TestPromptForConfigAsCode_UsingCacWithProjectStorage(t *testing.T) {
-	pa := []*testutil.PA{
-		testutil.NewConfirmPrompt("Would you like to use Config as Code?", "", true),
-		testutil.NewSelectPrompt("Select where to store the Git credentials", "", []string{"Library", "Project"}, "Project"),
-		testutil.NewInputPrompt("Git URL", "The URL of the Git repository to store configuration.", "https://github.com/blah.git"),
-		testutil.NewInputPromptWithDefault("Git repository base path", "The path in the repository where Config As Code settings are stored. Default value is '.octopus/'.", ".octopus/", "./octopus/project"),
-		testutil.NewInputPromptWithDefault("Git branch", "The default branch to use. Default value is 'main'.", "main", "main"),
-		testutil.NewConfirmPromptWithDefault("Is the 'main' branch protected?", "If the default branch is protected, you may not have permission to push to it.", false, false),
-		testutil.NewInputPrompt("Enter a protected branch pattern (enter blank to end)", "This setting only applies within Octopus and will not affect your protected branches in Git. Use wildcard syntax to specify the range of branches to include. Multiple patterns can be supplied", "test"),
-		testutil.NewInputPrompt("Enter a protected branch pattern (enter blank to end)", "This setting only applies within Octopus and will not affect your protected branches in Git. Use wildcard syntax to specify the range of branches to include. Multiple patterns can be supplied", ""),
-		testutil.NewInputPromptWithDefault("Initial Git commit message", "The commit message used in initializing. Default value is 'Initial commit of deployment process'.", "Initial commit of deployment process", "init message"),
-		testutil.NewInputPrompt("Git username", "The Git username.", "user1"),
-		testutil.NewPasswordPrompt("Git password", "The Git password.", "password"),
-	}
-
-	asker, checkRemainingPrompts := testutil.NewMockAsker(t, pa)
-
-	flags := projectCreate.NewCreateFlags()
-	flags.ConfigAsCode.Value = false
-	opts := projectCreate.NewCreateOptions(flags, &cmd.Dependencies{Ask: asker})
-	err := projectCreate.PromptForConfigAsCode(opts, nil)
-	checkRemainingPrompts()
-	assert.NoError(t, err)
-	assert.True(t, opts.ConfigAsCode.Value)
-	assert.Equal(t, "project", opts.GitStorage.Value)
-	assert.Equal(t, "https://github.com/blah.git", opts.GitUrl.Value)
-	assert.Equal(t, "./octopus/project", opts.GitBasePath.Value)
-	assert.Equal(t, "main", opts.GitBranch.Value)
-	assert.False(t, opts.GitDefaultBranchProtected.Value)
-	assert.Equal(t, "init message", opts.GitInitialCommitMessage.Value)
-	assert.Equal(t, []string{"test"}, opts.GitProtectedBranchPatterns.Value)
-	assert.Equal(t, "user1", opts.GitUsername.Value)
-	assert.Equal(t, "password", opts.GitPassword.Value)
-}
-
-func TestPromptForConfigAsCode_UsingCacWithLibraryStorage(t *testing.T) {
-	pa := []*testutil.PA{
-		testutil.NewConfirmPrompt("Would you like to use Config as Code?", "", true),
-		testutil.NewSelectPrompt("Select where to store the Git credentials", "", []string{"Library", "Project"}, "Library"),
-		testutil.NewInputPrompt("Git URL", "The URL of the Git repository to store configuration.", "https://github.com/blah.git"),
-		testutil.NewInputPromptWithDefault("Git repository base path", "The path in the repository where Config As Code settings are stored. Default value is '.octopus/'.", ".octopus/", "./octopus/project"),
-		testutil.NewInputPromptWithDefault("Git branch", "The default branch to use. Default value is 'main'.", "main", "main"),
-		testutil.NewConfirmPromptWithDefault("Is the 'main' branch protected?", "If the default branch is protected, you may not have permission to push to it.", false, false),
-		testutil.NewInputPrompt("Enter a protected branch pattern (enter blank to end)", "This setting only applies within Octopus and will not affect your protected branches in Git. Use wildcard syntax to specify the range of branches to include. Multiple patterns can be supplied", ""),
-		testutil.NewInputPromptWithDefault("Initial Git commit message", "The commit message used in initializing. Default value is 'Initial commit of deployment process'.", "Initial commit of deployment process", "init message"),
-		testutil.NewSelectPrompt("Select which Git credentials to use", "", []string{"Git Creds 1", "Git Creds 2"}, "Git Creds 2"),
-	}
-
-	asker, checkRemainingPrompts := testutil.NewMockAsker(t, pa)
-
-	gitCredsCallbackWasCalled := false
-	getGitCredentials := func() ([]*credentials.Resource, error) {
-		gitCredsCallbackWasCalled = true
-		creds := credentials.NewResource("Git Creds 1", credentials.NewReference("gitcreds-1"))
-		creds.ID = "gitcreds-1"
-		creds2 := credentials.NewResource("Git Creds 2", credentials.NewReference("gitcreds-2"))
-		creds2.ID = "gitcreds-2"
-		return []*credentials.Resource{creds, creds2}, nil
-	}
-	flags := projectCreate.NewCreateFlags()
-	flags.ConfigAsCode.Value = false
-	opts := projectCreate.NewCreateOptions(flags, &cmd.Dependencies{Ask: asker})
-	err := projectCreate.PromptForConfigAsCode(opts, getGitCredentials)
-	checkRemainingPrompts()
-	assert.NoError(t, err)
-	assert.True(t, opts.ConfigAsCode.Value)
-	assert.Equal(t, "library", opts.GitStorage.Value)
-	assert.Equal(t, "https://github.com/blah.git", opts.GitUrl.Value)
-	assert.Equal(t, "./octopus/project", opts.GitBasePath.Value)
-	assert.Equal(t, "main", opts.GitBranch.Value)
-	assert.Equal(t, "init message", opts.GitInitialCommitMessage.Value)
-	assert.Equal(t, "Git Creds 2", opts.GitCredentials.Value)
-	assert.True(t, gitCredsCallbackWasCalled)
-	assert.Empty(t, opts.GitUsername.Value)
-	assert.Empty(t, opts.GitPassword.Value)
-}
-
-func TestPromptForConfigAsCode_UsingCacWithBranchProtection(t *testing.T) {
-	pa := []*testutil.PA{
-		testutil.NewConfirmPrompt("Would you like to use Config as Code?", "", true),
-		testutil.NewSelectPrompt("Select where to store the Git credentials", "", []string{"Library", "Project"}, "Library"),
-		testutil.NewInputPrompt("Git URL", "The URL of the Git repository to store configuration.", "https://github.com/blah.git"),
-		testutil.NewInputPromptWithDefault("Git repository base path", "The path in the repository where Config As Code settings are stored. Default value is '.octopus/'.", ".octopus/", "./octopus/project"),
-		testutil.NewInputPromptWithDefault("Git branch", "The default branch to use. Default value is 'main'.", "main", "main"),
-		testutil.NewConfirmPromptWithDefault("Is the 'main' branch protected?", "If the default branch is protected, you may not have permission to push to it.", true, false),
-		testutil.NewInputPrompt("Initial commit branch name", "The branch where the Config As Code settings will be initially committed", "initial-branch"),
-		testutil.NewInputPrompt("Enter a protected branch pattern (enter blank to end)", "This setting only applies within Octopus and will not affect your protected branches in Git. Use wildcard syntax to specify the range of branches to include. Multiple patterns can be supplied", "test"),
-		testutil.NewInputPrompt("Enter a protected branch pattern (enter blank to end)", "This setting only applies within Octopus and will not affect your protected branches in Git. Use wildcard syntax to specify the range of branches to include. Multiple patterns can be supplied", ""),
-		testutil.NewInputPromptWithDefault("Initial Git commit message", "The commit message used in initializing. Default value is 'Initial commit of deployment process'.", "Initial commit of deployment process", "init message"),
-		testutil.NewSelectPrompt("Select which Git credentials to use", "", []string{"Git Creds 1", "Git Creds 2"}, "Git Creds 2"),
-	}
-
-	//pa := []*testutil.PA{
-	//	testutil.NewConfirmPrompt("Would you like to use Config as Code?", "", true),
-	//	testutil.NewSelectPrompt("Select where to store the Git credentials", "", []string{"Library", "Project"}, "Library"),
-	//	testutil.NewInputPrompt("Git URL", "The URL of the Git repository to store configuration.", "https://github.com/blah.git"),
-	//	testutil.NewInputPrompt("Git repository base path", "The path in the repository where Config As Code settings are stored. Default value is '.octopus/'.", "./octopus/project"),
-	//	testutil.NewInputPromptWithDefault("Git branch", "The default branch to use. Default value is 'main'.", "main", "main"),
-	//	testutil.NewConfirmPromptWithDefault("Is the 'main' branch protected?", "If the default branch is protected, you may not have permission to push to it.", true, false),
-	//	testutil.NewInputPrompt("Initial commit branch", "The branch where the Config As Code settings will be initially committed", "initial-commit"),
-	//	testutil.NewInputPrompt("Initial Git commit message", "The commit message used in initializing. Default value is 'Initial commit of deployment process'.", "init message"),
-	//	testutil.NewSelectPrompt("Select which Git credentials to use", "", []string{"Git Creds 1", "Git Creds 2"}, "Git Creds 2"),
-	//}
-
-	asker, checkRemainingPrompts := testutil.NewMockAsker(t, pa)
-
-	gitCredsCallbackWasCalled := false
-	getGitCredentials := func() ([]*credentials.Resource, error) {
-		gitCredsCallbackWasCalled = true
-		creds := credentials.NewResource("Git Creds 1", credentials.NewReference("gitcreds-1"))
-		creds.ID = "gitcreds-1"
-		creds2 := credentials.NewResource("Git Creds 2", credentials.NewReference("gitcreds-2"))
-		creds2.ID = "gitcreds-2"
-		return []*credentials.Resource{creds, creds2}, nil
-	}
-	flags := projectCreate.NewCreateFlags()
-	flags.ConfigAsCode.Value = false
-
-	opts := projectCreate.NewCreateOptions(flags, &cmd.Dependencies{Ask: asker})
-	err := projectCreate.PromptForConfigAsCode(opts, getGitCredentials)
-	checkRemainingPrompts()
-	assert.NoError(t, err)
-	assert.True(t, opts.ConfigAsCode.Value)
-	assert.Equal(t, "library", opts.GitStorage.Value)
-	assert.Equal(t, "https://github.com/blah.git", opts.GitUrl.Value)
-	assert.Equal(t, "./octopus/project", opts.GitBasePath.Value)
-	assert.Equal(t, "main", opts.GitBranch.Value)
-	assert.Equal(t, "init message", opts.GitInitialCommitMessage.Value)
-	assert.Equal(t, "Git Creds 2", opts.GitCredentials.Value)
-	assert.True(t, opts.GitDefaultBranchProtected.Value)
-	assert.Equal(t, "initial-branch", opts.GitInitialCommitBranch.Value)
-
-	assert.True(t, gitCredsCallbackWasCalled)
-	assert.Empty(t, opts.GitUsername.Value)
-	assert.Empty(t, opts.GitPassword.Value)
 }
